% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{fetch.CurrencyRange}
\alias{fetch.CurrencyRange}
\title{Fetches currency conversion data}
\usage{
\method{fetch}{CurrencyRange}(x)
}
\arguments{
\item{x}{currency range object}
}
\value{
A currency data object
}
\description{
Fetches currency conversion data
}
\examples{
fetch(rng)
}
