% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{fetch_cur_data}
\alias{fetch_cur_data}
\title{Fetches time series with conversion rates for two currencies and year count}
\usage{
fetch_cur_data(cur_from = "CHF", cur_to = "RUB", years = 1)
}
\arguments{
\item{cur_from}{string (3-letter ISO Code) source currency}

\item{cur_to}{string (3-letter ISO code) target currency}

\item{years}{a number of years}
}
\value{
A data frame
}
\description{
Fetches time series with conversion rates for two currencies and year count
}
\examples{
fetch_cur_data()
fetch_cur_data(from_cur="USD")
}
