% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{make_cur_range}
\alias{make_cur_range}
\title{Creates a currency range object}
\usage{
make_cur_range(
  from_cur = "CHF",
  to_cur = "RUB",
  day = lubridate::day(Sys.Date()),
  month = lubridate::month(Sys.Date()),
  year_from = lubridate::year(Sys.Date()) - 1,
  year_to = lubridate::year(Sys.Date())
)
}
\arguments{
\item{from_cur}{string (3-letter ISO Code) from which currency to convert}

\item{to_cur}{string (3-letter ISO code) to which currency to convert}

\item{day}{string (3-letter ISO code) to which currency to convert}

\item{month}{string (3-letter ISO code) to which currency to convert}

\item{year_from}{from which year to convert}

\item{year_to}{to which year to convert}
}
\value{
A currency range object
}
\description{
Creates a currency range object
}
\examples{
make_cur_range()
make_cur_range(from_cur="USD")
}
