% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{plot.CurrencyData}
\alias{plot.CurrencyData}
\title{Plots currency conversion data}
\usage{
\method{plot}{CurrencyData}(x)
}
\arguments{
\item{x}{currency data object}
}
\value{
A ggplot2 plot of time series
}
\description{
Plots currency conversion data
}
\examples{
plot(cur_data)
}
